refactor(lowy): move daemon-startup primitives out of the supervisor
tryAcquirePidFile + daemonExecArgv are daemon-side launcher primitives
(imported by agent/main.ts) but lived in supervisor.ts alongside the
kolu-server-facing connect/spawn/version API — a discovery hazard, and a
different volatility axis (pid-file strategy / exec-flag policy vs how
kolu-server talks to a running daemon). Move both to daemon/daemonUtils.ts;
supervisor.ts and agent/main.ts import from there.
