+++ packages/server/src/daemon/supervisor.ts
@@ -31,19 +31,10 @@
  */
 
 import { spawn as spawnChild } from "node:child_process";
-import {
-  closeSync,
-  existsSync,
-  mkdirSync,
-  openSync,
-  readFileSync,
-  unlinkSync,
-  writeSync,
-} from "node:fs";
+import { existsSync, openSync } from "node:fs";
 import { createConnection, type Socket } from "node:net";
-import { dirname } from "node:path";
-import type { ClientRetryPluginContext } from "@orpc/client/plugins";
 import { createStdioCellsClient } from "@kolu/surface/links/stdio";
+import type { ClientRetryPluginContext } from "@orpc/client/plugins";
 import type { ContractRouterClient } from "@orpc/contract";
 import {
   AGENT_CONTRACT_VERSION,
@@ -53,6 +44,7 @@ import {
 import pkg from "../../package.json" with { type: "json" };
 import { daemonPaths } from "../koluState.ts";
 import { log } from "../log.ts";
+import { daemonExecArgv } from "./daemonUtils.ts";
 
 /** The typed client for the daemon's `agentSurface` contract, with the
  *  `ClientRetryPlugin` context threaded through (the same shape
@@ -75,83 +67,6 @@ export interface DaemonHandle {
   state(): "live" | "closed";
   /** Close the supervisor's socket (does NOT kill the daemon). */
   dispose(): void;
-}
-
-/** Atomically claim the pid file. Returns true if this process is the
- *  authoritative owner; false if another live daemon already owns it.
- *  Stale pid files (recorded pid no longer alive) are cleaned up and the
- *  gate is retried — keeps the gate working across crashes without
- *  external cleanup.
- *
- *  Shared by the daemon entrypoint (`../agent/main.ts`), which calls this
- *  before binding the socket so only one daemon per `$KOLU_STATE_DIR`
- *  ever owns the PTYs. */
-export function tryAcquirePidFile(pidFile: string): boolean {
-  mkdirSync(dirname(pidFile), { recursive: true, mode: 0o700 });
-  for (let attempt = 0; attempt < 4; attempt++) {
-    try {
-      const fd = openSync(pidFile, "wx", 0o600);
-      writeSync(fd, `${process.pid}\n`);
-      closeSync(fd);
-      return true;
-    } catch (err) {
-      if ((err as NodeJS.ErrnoException).code !== "EEXIST") throw err;
-    }
-    let recordedPid = 0;
-    try {
-      recordedPid = Number.parseInt(readFileSync(pidFile, "utf8").trim(), 10);
-    } catch {
-      // Unreadable; treat as stale.
-    }
-    if (Number.isFinite(recordedPid) && recordedPid > 0) {
-      try {
-        process.kill(recordedPid, 0);
-        return false; // Another daemon is alive.
-      } catch (sigErr) {
-        if ((sigErr as NodeJS.ErrnoException).code !== "ESRCH") throw sigErr;
-        // Stale — owner is gone. Fall through to unlink + retry.
-      }
-    }
-    try {
-      unlinkSync(pidFile);
-    } catch {
-      // Race: another process unlinked it first; loop and retry create.
-    }
-  }
-  // Couldn't claim after several attempts — treat as already-owned.
-  return false;
-}
-
-/** node exec flags the daemon must NOT inherit. `--watch` would make the
- *  detached daemon restart on source edits — killing every PTY mid-dev,
- *  the exact opposite of R-4's point. `--inspect*` would make it try to
- *  bind this process's debug port and fail to start. We keep only the
- *  loader/import flags that let node run the TS entry. */
-const DROP_EXEC_FLAG =
-  /^--(watch|watch-path|watch-preserve-output|inspect|inspect-brk|inspect-port|inspect-wait|debug|debug-brk)\b/;
-
-/** Strip dev-only flags from `process.execArgv`, preserving each kept
- *  flag's space-separated value (e.g. `--import tsx`). Exported for unit
- *  testing the filter in isolation. */
-export function daemonExecArgv(execArgv: string[]): string[] {
-  const out: string[] = [];
-  for (let i = 0; i < execArgv.length; i++) {
-    const tok = execArgv[i] as string;
-    if (!tok.startsWith("-")) {
-      out.push(tok);
-      continue;
-    }
-    // A following non-flag token is this flag's value (`--import tsx`).
-    const next = execArgv[i + 1];
-    const hasValue =
-      !tok.includes("=") && next !== undefined && !next.startsWith("-");
-    if (!DROP_EXEC_FLAG.test(tok)) {
-      out.push(tok);
-      if (hasValue) out.push(next as string);
-    }
-    if (hasValue) i++; // consume the value regardless of keep/drop
-  }
-  return out;
 }
 
 /** Attempt a single connect with a short timeout. Returns the socket on
